Add support for customizing the serialized type name
This commit keeps the default behavior, but adds the ability to provide
an overriding Func<Type,string> method that generates the string
representing the type. Json.NET exposes this behavior using the
SerializationBinder class in the BCL, however only in .NET 4 does that
class support two way binding. And on the plus side, a Func is cleaner
and easier to implement.