Implmented tests and fixes for quoted properties and properties with comma in querystrings
Properties that had values that where quoted never kept their quotes when parsed
In jsv quotes have to be escaped with another quoted, this was not done for querystrings

Properties that had valus containing commas caused problems when parsing lists from
the querystring. By quoting these values the comma is parsed corectly on the way into
a service endpoint.