Fix a Null Pointer Exception on Deserialization
When deserializing a collection and a property of a serialzed collection member has an empty value in the jsv file, the member can not be deserialized because of a null pointer exception.
The cause for that was found in the EatValue(..) method of JsvTypeSerializer: <null> was returned instead of an empty string at lines 380 and 391. That later caused a null exception when e.g. the UnescapeString(..) method of JsvTypeSerializer is called.