JsonSerializer/JsvTypeSerializer inconsistently uses the AssumeUtc config setting
DateTime handling is inconsistent with respect to the AssumcUtc when using the JSConfig setting.

If serializing to JSON - The JsonTypeSerializer will use the UTC specification type and convert Unspecified DateTime values to UTC in method "WriteWcfJsonDate"

If serializing to CSV- The JsvTypeSerializer will not use the UTC specification type and does not convert Unspecified DateTime values to UTC in method "ToShortestXsdDateTimeString"

Adding an internal extension method "UseConfigSpecifiedSetting" to match DateTime conversion between the two different serializers.